Move getUpstreamProviderOrg to steps.go
